Switch to factory interface instead of exporting class
