+++ addons/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    BaseStore: require('./BaseStore'),
+    createStore: require('./createStore')
+};
