+++ docs/dispatchr.md
@@ -1,18 +1,22 @@
-# Dispatcher API
+# Dispatchr API
 
-## Constructor(context)
+Dispatchr has one main function that is exported: `createDispatcher(options)`. This returns a new [`Dispatcher`](#dispatcher-api) instance. `createDispatcher` takes the following `options`:
 
-Creates a new Dispatcher instance with the following parameters:
+ * `options.stores`: An array of stores to register automatically
 
- * `context`: A context object that will be made available to all stores. Useful for request or session level settings.
-
-## Static Methods
+## Dispatcher API
 
 ### registerStore(storeClass)
 
-A static method to register stores to the Dispatcher class making them available to handle actions and be accessible through `getStore` on Dispatchr instances.
+A static method to register stores to the dispatcher making them available to handle actions and be accessible through `getStore` on the dispatcher context.
 
-## Instance Methods
+### createContext(contextOptions)
+
+Creates a new dispatcher [context](#context-api) that isolates stores and dispatches with the following parameters:
+
+ * `contextOptions`: A context object that will be made available to all stores. Useful for request or session level settings.
+
+## Context API
 
 ### dispatch(actionName, payload)
 
@@ -21,27 +25,27 @@ Dispatches an action, in turn calling all stores that have registered to handle 
  * `actionName`: The name of the action to handle (should map to store action handlers)
  * `payload`: An object containing action information.
 
-### getStore(storeClass)
+#### getStore(storeClass)
 
 Retrieve a store instance by class. Allows access to stores from components or stores from other stores.
 
 ```js
-var store = require('./stores/MessageStore');
-dispatcher.getStore(store);
+var MessageStore = require('./stores/MessageStore');
+dispatcher.getStore(MessageStore);
 ```
 
-### waitFor(storeClasses, callback)
+#### waitFor(storeClasses, callback)
 
 Waits for another store's handler to finish before calling the callback. This is useful from within stores if they need to wait for other stores to finish first.
 
   * `storeClasses`: An array of store classes to wait for
   * `callback`: Called after all stores have fully handled the action
 
-### dehydrate()
+#### dehydrate()
 
-Returns a serializable object containing the state of the Dispatchr instance as well as all stores that have been used since instantiation. This is useful for serializing the state of the application to send it to the client.
+Returns a serializable object containing the state of the dispatcher context as well as all stores that have been used since instantiation. This is useful for serializing the state of the application to send it to the client.
 
-### rehydrate(dispatcherState)
+#### rehydrate(dispatcherState)
 
-Takes an object representing the state of the Dispatchr instance (usually retrieved from dehydrate) to rehydrate the instance as well as the store instance state.
+Takes an object representing the state of the dispatcher context (usually retrieved from dehydrate) to rehydrate the instance as well as the store instance state.
