+++ tests/mock/NoDehydrate.js
@@ -2,7 +2,7 @@
  * Copyright 2014, Yahoo! Inc.
  * Copyrights licensed under the New BSD License. See the accompanying LICENSE file for terms.
  */
-var createStore = require('../../utils/createStore');
+var createStore = require('../../addons/createStore');
 
 module.exports = createStore({
     storeName: 'NoDehydrateStore',
