+++ utils/index.js
@@ -1,0 +1,1 @@
+throw new Error("require('fluxible/utils') has moved to require('fluxible/addons')");
